Added a C++ Version
Added a C++ version of the recolor inverter program. C++ is more similar to other bianary compiled languages so it should be easy to translate into C, C#, Java, you name it. This program was programmed by me and one function was made by JDJG. This is based off of JDJG's python recolor Converter.